Save the charms given by Chaos Orb in the savefile
This makes it so loading a save where Chaos Orb was equipped
after launching the game will keep the same charms being given
instead of starting with nothing and having to wait for a reroll.
It also makes for more sensible behaviour when switching between
multiple saves with Chaos Orb equipped; before the set of Orb-given
charms would transfer over from one to the other.