Fix infinite loop when generating a seed if Add Charms is off but charms are not randomized
The maximum charm cost would still be raised above 40, but if a Salubra check
required more than 40 charms, there would not be enough of them in the pool
to put it into logic.

A better solution to this is to add the charm vanilla locations, along with
logic for them, so that rando can include the Transcendence charms in logic
even if charms aren't randomized.