Fix Fungus2_25 formerly-oneway transition not being available when entering the room from the right
This happened because the OnSceneChange hook runs with the old scene still
loaded, and so GameObject.Find would look for the requested object in the old
scene first. This did not matter in most cases, but Fungus2_15, the room to
the right of Fungus2_25, does have its own top2 transition, which was being
found instead of the one we wanted.